Schema validation working
