+++ app/middlewares/user.js
@@ -1,7 +1,4 @@
-const { validationResult, checkSchema } = require('express-validator');
-
-const { emailSchema, passwordSchema } = require('../schemas/user');
-const { modelValidationError, emailRepeatedError } = require('../errors');
+const { emailRepeatedError } = require('../errors');
 const { findUserByEmail } = require('../services/user');
 const logger = require('../logger');
 
@@ -15,31 +12,3 @@ exports.checkMailIsAlreadyInUse = ({ body: { email } }, _response, next) => {
     }
   });
 };
-
-exports.checkUserSchema = () => [
-  checkSchema({
-    email: emailSchema,
-    password: passwordSchema,
-    firstName: {
-      exists: {
-        errorMessage: 'First name must be present'
-      }
-    },
-    lastName: {
-      exists: {
-        errorMessage: 'Last name must be present'
-      }
-    }
-  }),
-  (request, _response, next) => {
-    const errors = validationResult(request);
-    if (errors.isEmpty()) {
-      next();
-    } else {
-      const errorMessage = errors.map(error => error.msg).join('; ');
-
-      logger.error(`Invalid user params: ${errorMessage}`);
-      next(modelValidationError(errorMessage));
-    }
-  }
-];
